Remove deprecated convention based invocationConfig loading
Updated all test fixtures to have a `src/index` file that exposes
the invocationConfig.

Updated the CHANGELOG to reflect the changes.
Added an entry to the CHANGELOG to reflect a change that was
made to the `integration-sdk-testing` package around changes to
convention based loading as well.

Added missing devDependencies to the `integration-sdk-cli` package.