Support for omitting specific graph objects from file storage
Some entities and relationships do not need to be stored on the
file system at all. We only need to store graph objects on the file
system if we later intend to fetch the data from the job state or
iterate over the `_type`. This PR introduces support for specifying
metadata in each step that can be used by the `FileSystemGraphObjectStore`
that will be used to omit specific data from being written to disk
entirely. The data will still get uploaded.