INT-5226 Add log statement if an error occurs POSTing with axios
We are seeing ERR_INVALID_ARG_TYPE errors that seem to be coming from
axios. However, we are losing a lot of helpful data about the error
when we handle it later in the lifecycle when we throw our custom
IntegrationError. While we investigate these errors, add a log statement
to get more details on why axios threw.