Fix #823 - Do not modify `JUPITERONE_API_KEY` env var in CLI defaults
The `j1-integration run` command accepts at `--api-key` CLI flag.
Previously, this CLI flag had a default value that would read the
value of the `JUPITERONE_API_KEY` environment variable and modify
the value using a regex replace. The modification of the value of
`JUPITERONE_API_KEY` resulted in an invalid value being supplied
in request headers. The JupiterOne API was responding with 401
Unauthorized consistently.

After discussing with the team, it turns out that the regex
replacement on `JUPITERONE_API_KEY` was put in place to prevent
the value of `JUPITERONE_API_KEY` from being logged to the
console in the CLI help text message.

The usage of the `commander` CLI options has been updated to
include specific help text messages instead of just logging the
current value of each respective environment variable.